Update README.md
Notations on the effects of not having red alert collaborators and others involved in the manner they are safe rather than other premium services where they conversate one one one. Its best to be in your group and identified as a member of one. Even note if you are identified as external.